requirements.txt
# Dependencias para el Proyecto de Reconocimiento de Señales de Tráfico

# Framework de Deep Learning
tensorflow>=2.13.0

# Descarga de datasets desde Kaggle
kagglehub>=0.2.0

# Aplicación Web
streamlit>=1.28.0

# Procesamiento de Imágenes
Pillow>=10.0.0

# Computación Numérica
numpy>=1.24.0

# Visualización de Datos
matplotlib>=3.7.0

# Procesamiento de datos
pandas>=2.0.0
scikit-learn>=1.3.0

# Nota: Si tienes GPU NVIDIA, puedes instalar tensorflow-gpu en lugar de tensorflow
# para acelerar el entrenamiento significativamente.